+++ linux2mqtt/type_definitions.py
@@ -1,6 +1,8 @@
 """linux2mqtt type definitions."""
 
-from typing import Literal, TypedDict
+from typing import Literal, NotRequired, TypedDict
+
+Addr = tuple[str, int]
 
 StatusType = Literal["online", "offline"]
 """Metric status"""
@@ -18,6 +20,8 @@ class Linux2MqttConfig(TypedDict):
         Log verbosity
     homeassistant_prefix
         MQTT discovery topic prefix
+    homeassistant_disable_attributes
+        Disable attributes in home assistant discovery and exposes everything as entities
     linux2mqtt_hostname
         A descriptive name for the system being monitored
     mqtt_client_id
@@ -43,6 +47,7 @@ class Linux2MqttConfig(TypedDict):
 
     log_level: str
     homeassistant_prefix: str
+    homeassistant_disable_attributes: bool
     linux2mqtt_hostname: str
     mqtt_client_id: str
     mqtt_user: str
@@ -66,12 +71,65 @@ class LinuxDeviceEntry(TypedDict):
         The name of the device to display in home assistant
     model
         The model of the device as additional info
+    hw_version
+        The hardware version of the device (OS version)
+    sw_version
+        The software version of the device (Linux2mqtt version)
 
     """
 
     identifiers: str
     name: str
     model: str
+    hw_version: NotRequired[str]
+    sw_version: NotRequired[str]
+
+
+class MetricEntities(TypedDict):
+    """A metric entity object for discovery in home assistant.
+
+    Attributes
+    ----------
+    name
+        The name of the sensor to display in home assistant
+    state_field
+        The field in the state topic to extract the state value from
+    icon
+        The icon of the sensor to display
+    unit_of_measurement
+        The unit of measurement of the sensor
+    device_class
+        The device class of the sensor
+
+    """
+
+    name: str
+    state_field: str
+    icon: str | None
+    unit_of_measurement: str | None
+    device_class: str | None
+
+
+class AvailabilityEntry(TypedDict):
+    """An availability entity object for discovery in home assistant.
+
+    Attributes
+    ----------
+    topic
+        The MQTT topic receiving availability updates
+    value_template
+        Template to extract device's availability from the topic
+    payload_available
+        The payload that represents the available state
+    payload_not_available
+        The payload that represents the unavailable state
+
+    """
+
+    topic: str
+    value_template: NotRequired[str]
+    payload_available: NotRequired[str]
+    payload_not_available: NotRequired[str]
 
 
 class LinuxEntry(TypedDict):
@@ -85,8 +143,10 @@ class LinuxEntry(TypedDict):
         The unique id of the sensor in home assistant
     icon
         The icon of the sensor to display
-    availability_topic
-        The topic to check the availability of the sensor
+    availability
+        The list of topics to check the availability of the sensor
+    availability_mode
+        The conditions needed to set the entity to available
     payload_available
         The payload of availability_topic of the sensor when available
     payload_unavailable
@@ -105,7 +165,7 @@ class LinuxEntry(TypedDict):
         The device the sensor is attributed to
     device_class
         The device class of the sensor
-    state_topic
+    json_attributes_topic
         The topic containing all information for the attributes of the sensor
     qos
         The QOS of the discovery message
@@ -115,7 +175,8 @@ class LinuxEntry(TypedDict):
     name: str
     unique_id: str
     icon: str | None
-    availability_topic: str
+    availability: list[AvailabilityEntry]
+    availability_mode: NotRequired[str]
     payload_available: str
     payload_not_available: str
     state_topic: str
@@ -125,5 +186,5 @@ class LinuxEntry(TypedDict):
     payload_off: str
     device: LinuxDeviceEntry
     device_class: str | None
-    json_attributes_topic: str
+    json_attributes_topic: str | None
     qos: int